example for SMART

+++ elementary/BaseART.py
@@ -91,7 +91,7 @@ class BaseART(BaseEstimator, ClusterMixin):
         self.check_dimensions(X)
 
         self.W: list[np.ndarray] = []
-        self.labels_ = np.zeros((X.shape[0], ))
+        self.labels_ = np.zeros((X.shape[0], ), dtype=int)
         for _ in range(max_iter):
             for i, x in enumerate(X):
                 self.pre_step_fit(X)
